### 2022-02-14 #4159 @danielskatz
From the submitters

> Our submission is about:
> * `scipy.stats.tukey_hsd`, a function that performs Tukey's honestly significant difference (HSD) test for pairwise comparisons of observations from several groups and
> * `scipy.stats.studentized_range`, a class representing the studentized range distribution, a statistical distribution used in Tukey's HSD test and other post-hoc tests.
>
> `tukey_hsd` and `studentized_range` were contributed to the SciPy library and are included in SciPy's latest release, 1.8.0.
>
> We described the development and validation of `studentized_range` in a "virtual poster", which took the form of a Jupyter notebook, at the SciPy 2021 conference. While the submission was "reviewed by peers" in a sense, this venue differs from typical peer-reviewed publications in that 1) the submission was accepted based only on an abstract, and 2) the reviewers were not selected based on their expertise in the field.
>
> The URL submitted here is to a fork of the SciPy library. The branch containing the paper is "joss_studentized_range" Since the contributions are parts of a much larger library, the following links may be helpful:
> - `tukey_hsd` pull request: https://github.com/scipy/scipy/pull/13002
> - `tukey_hsd` documentation: http://scipy.github.io/devdocs/reference/generated/scipy.stats.tukey_hsd.html#scipy.stats.tukey_hsd
> - `tukey_hsd` source code: https://github.com/scipy/scipy/blob/5d0c1b8c55cb16c25060772966c0d179beea3ccf/scipy/stats/_hypotests.py#L2201-L2382
> - ` studentized_range` pull request: https://github.com/scipy/scipy/pull/13732
> - ` studentized_range` documentation: http://scipy.github.io/devdocs/reference/generated/scipy.stats.studentized_range.html#scipy.stats.studentized_range
> - ` studentized_range` source code: https://github.com/scipy/scipy/blob/5d0c1b8c55cb16c25060772966c0d179beea3ccf/scipy/stats/_continuous_distns.py#L8957-L9141
