### 2022-02-14 #4159 @danielskatz
👋 @mdhaber - Please work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch joss_studentized_range` to check again, and the command `@whedon generate pdf from branch joss_studentized_range` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
