### 2022-02-14 #4159 @danielskatz
@mdhaber -  Due to the relatively small amount of code in (https://github.com/scipy/scipy/blob/5d0c1b8c55cb16c25060772966c0d179beea3ccf/scipy/stats/_hypotests.py#L2201-L2382 & https://github.com/scipy/scipy/blob/5d0c1b8c55cb16c25060772966c0d179beea3ccf/scipy/stats/_continuous_distns.py#L8957-L9141, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.
