### 2022-02-17 45a7582
Add DOIs suggested by whedon - thanks whedon!
