+++ paper.bib
@@ -4,7 +4,8 @@
   journal={Biometrics},
   pages={99--114},
   year={1949},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/3001913}
 }
 
 @article{lund1983algorithm,
@@ -15,7 +16,8 @@
   number={2},
   pages={204--210},
   year={1983},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/2347300}
 }
 
 @article{virtanen2020scipy,
@@ -26,7 +28,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
 }
 
 @misc{PRtukey_hsd,
@@ -45,7 +48,8 @@
   volume={57},
   pages={61},
   year={2010},
-  organization={Austin, TX}
+  organization={Austin, TX},
+  doi={10.25080/majora-92bf1922-011}
 }
 
 @article{vallat2018pingouin,
@@ -55,7 +59,8 @@
   volume={3},
   number={31},
   pages={1026},
-  year={2018}
+  year={2018},
+  doi={10.21105/joss.01026}
 }
 
 @misc{PRstudentized_range,
@@ -72,5 +77,6 @@
   author={Samuel Wallan and Dominic Chmiel and Matt Haberland},
   booktitle={Proceedings of the 9th Python in Science Conference},
   year={2021},
-  url={https://doi.org/10.25080/majora-1b6fd038-01e}
+  url={https://doi.org/10.25080/majora-1b6fd038-01e},
+  doi={10.25080/majora-1b6fd038-01e}
 }
