### 2022-02-24 361d241
ENH: stats: update gumbel_r  fit
