### 2022-02-24 9d19853
Update scipy/stats/_continuous_distns.py
