### 2022-02-26 97248a4
DOC: remove py:obj references to matplotlib [skip actions] [skip azp]
