### 2022-03-03 #4159 @mikldk
@mdhaber - thank you for the submission.

A few minor requests:

- [ ] Can you please add country to your affiliation?
- [ ] Please write footnotes with capital first letter

Do you have any suggestions for reviewers ([here](https://bit.ly/joss-reviewers) is a list of potential reviewers) - please do not tag (`@`) anyone yet?
