### 2022-03-03 2b0471d
Capitalize first letter of footnotes, add country to affiliation
