### 2022-03-03 #4159 @mdhaber
Thanks for editing @mikldk, and sure thing.

The footnotes were "co-first author" and "corresponding author", right? If so, perhaps the example [here](https://joss.readthedocs.io/en/latest/submitting.html) could be updated with those suggestions, too? LMK if you'd like me to submit a PR!

These reviewers seem appropriate due to their interest in statistics and statistical/scientific computing with Python as a preferred language: tylerwmarrs, ali-ramadhan
and those with Python as an "Other Programming Language": chlalanne, mcavs, matbesancon
Also, with interest in statistics and Python as a preferred language: acolum, stensy, stulacy, Nazeeefa, tleonardi, dsurujon, No
and those with Python as an "Other Programming Language": Jminnier, nhejazi, franjoim, cpalmer718, tomfaulkenberry, jean997
