### 2022-03-05 #4159 @arfon
> The footnotes were "co-first author" and "corresponding author", right? If so, perhaps the example [here](https://joss.readthedocs.io/en/latest/submitting.html) could be updated with those suggestions, too? LMK if you'd like me to submit a PR!

We should probably fix those (PRs welcome!). I'd prefer that we use a capital first letter too.
