+++ scipy/stats/_stats_py.py
@@ -53,6 +53,7 @@ from ._hypotests import _all_partitions, _batch_generator
 from ._hypotests_pythran import _compute_outer_prob_inside_method
 from ._axis_nan_policy import (_axis_nan_policy_factory,
                                _broadcast_concatenate)
+from ._binomtest import _binary_search_for_binom_tst as _binary_search
 from scipy._lib._bunch import _make_tuple_bunch
 
 
@@ -4422,46 +4423,50 @@ def fisher_exact(table, alternative='two-sided'):
     n2 = c[1, 0] + c[1, 1]
     n = c[0, 0] + c[1, 0]
 
-    def binary_search(n, n1, n2, side):
-        """Binary search for where to begin halves in two-sided test."""
-        if side == "upper":
-            minval = mode
-            maxval = n
-        else:
-            minval = 0
-            maxval = mode
-        guess = -1
-        while maxval - minval > 1:
-            if maxval == minval + 1 and guess == minval:
-                guess = maxval
-            else:
-                guess = (maxval + minval) // 2
-            pguess = hypergeom.pmf(guess, n1 + n2, n1, n)
-            if side == "upper":
-                ng = guess - 1
-            else:
-                ng = guess + 1
-            if pguess <= pexact < hypergeom.pmf(ng, n1 + n2, n1, n):
-                break
-            elif pguess < pexact:
-                maxval = guess
-            else:
-                minval = guess
-        if guess == -1:
-            guess = minval
-        if side == "upper":
-            while guess > 0 and \
-                    hypergeom.pmf(guess, n1 + n2, n1, n) < pexact * epsilon:
-                guess -= 1
-            while hypergeom.pmf(guess, n1 + n2, n1, n) > pexact / epsilon:
-                guess += 1
-        else:
-            while hypergeom.pmf(guess, n1 + n2, n1, n) < pexact * epsilon:
-                guess += 1
-            while guess > 0 and \
-                    hypergeom.pmf(guess, n1 + n2, n1, n) > pexact / epsilon:
-                guess -= 1
-        return guess
+    # def binary_search(n, n1, n2, side):
+    #     """Binary search for where to begin halves in two-sided test."""
+    #     if side == "upper":
+    #         minval = mode
+    #         maxval = n
+    #     else:
+    #         minval = 0
+    #         maxval = mode
+    #     guess = -1
+    #     while maxval - minval > 1:
+    #         print(guess, minval, maxval)
+    #         if maxval == minval + 1 and guess == minval:
+    #             guess = maxval
+    #         else:
+    #             guess = (maxval + minval) // 2
+    #         pguess = hypergeom.pmf(guess, n1 + n2, n1, n)
+    #         if side == "upper":
+    #             ng = guess - 1
+    #         else:
+    #             ng = guess + 1
+    #         if pguess <= pexact < hypergeom.pmf(ng, n1 + n2, n1, n):
+    #             break
+    #         elif pguess < pexact:
+    #             maxval = guess
+    #         else:
+    #             minval = guess
+    #     if guess == -1:
+    #         guess = minval
+    #     print('here')
+    #     if side == "upper":
+    #         while guess > 0 and \
+    #                 hypergeom.pmf(guess, n1 + n2, n1, n) < pexact * epsilon:
+    #             guess -= 1
+    #         while hypergeom.pmf(guess, n1 + n2, n1, n) > pexact / epsilon:
[…]
