### 2022-03-27 c5ebcb1
ENH: stats: add Pascal distribution
