### 2022-03-28 7b11b83
ENH: stats: add TruncatedDistribution
