+++ scipy/stats/__init__.py
@@ -484,6 +484,7 @@ from ._rvs_sampling import rvs_ratio_uniforms, NumericalInverseHermite  # noqa
 from ._page_trend_test import page_trend_test
 from ._mannwhitneyu import mannwhitneyu
 from ._fit import fit
+from ._distn_infrastructure import TruncatedDistribution
 
 # Deprecated namespaces, to be removed in v2.0.0
 from . import (
