+++ scipy/stats/_distn_infrastructure.py
@@ -4053,3 +4053,185 @@ def get_distribution_names(namespace_pairs, rv_base_class):
         if isinstance(value, rv_base_class):
             distn_names.append(name)
     return distn_names, distn_gen_names
+
+
+# TODO: how to override methods
+# TODO: x beyond support
+class TruncatedDistribution(rv_continuous):
+    def _parse_args(self, x, left, right, *args, **kwargs):
+        a, b = self.dist.support(*args, **kwargs)
+        if left is None:
+            left = a
+        if right is None:
+            right = b
+        shapes, loc, scale = self.dist._parse_args(*args, **kwargs)
+
+        all_arrays = (a, b, left, right, loc, scale) + (shapes)
+        if x is not None:
+            all_arrays = (x,) + all_arrays
+
+        params = np.broadcast_arrays(*all_arrays)
+
+        if x is None:
+            a, b, left, right, loc, scale = params[:6]
+            shapes = params[6:]
+        else:
+            x, a, b, left, right, loc, scale = params[:7]
+            shapes = params[7:]
+
+        i = a > left
+        left[i] = a[i]
+        i = b < right
+        right[i] = b[i]
+
+        # replace this
+        assert np.all(left < right)
+
+        return x, left, right, shapes, loc, scale
+
+    def _get_norm(self, left, right, *args, **kwargs):
+        return (self.dist.cdf(right, *args, **kwargs)
+                - self.dist.cdf(left, *args, **kwargs))
+
+    def _get_log_norm(self, left, right, *args, **kwargs):
+        # can we improve this?
+        return np.log(self._get_norm(left, right, *args, **kwargs))
+
+    def __init__(self, dist):
+        self.dist = dist
+        self.name = f"truncated_{dist.name}"
+
+    def support(self, *args, left=None, right=None, **kwargs):
+        params = self._parse_args(None, left, right, *args, **kwargs)
+        x, left, right, shapes, loc, scale = params
+
+        return left[()], right[()]
+
+    def pdf(self, x, *args, left=None, right=None, **kwargs):
+        params = self._parse_args(x, left, right, *args, **kwargs)
+        x, left, right, shapes, loc, scale = params
+
+        norm = self._get_norm(left, right, *args, **kwargs)
+
+        return self.dist.pdf(x, *args, **kwargs) / norm
+
+    def logpdf(self, x, *args, left=None, right=None, **kwargs):
+        params = self._parse_args(x, left, right, *args, **kwargs)
+        x, left, right, shapes, loc, scale = params
+
+        log_norm = self._get_log_norm(left, right, *args, **kwargs)
+
+        return self.dist.logpdf(x, *args, **kwargs) - log_norm
+
+    def cdf(self, x, *args, left=None, right=None, **kwargs):
+        params = self._parse_args(x, left, right, *args, **kwargs)
+        x, left, right, shapes, loc, scale = params
+
+        norm = self._get_norm(left, right, *args, **kwargs)
+
+        return (self.dist.cdf(x, *args, **kwargs)
+                - self.dist.cdf(left, *args, **kwargs))/norm
+
+    def logcdf(self, x, *args, left=None, right=None, **kwargs):
+        params = self._parse_args(x, left, right, *args, **kwargs)
+        x, left, right, shapes, loc, scale = params
+
+        log_norm = self._get_log_norm(left, right, *args, **kwargs)
+
+        # can we improve this?
+        return np.log(self.dist.cdf(x, *args, **kwargs)
+                      - self.dist.cdf(left, *args, **kwargs)) - log_norm
+
+    def sf(self, x, *args, left=None, right=None, **kwargs):
[…]
