### 2022-05-09 #4383 @editorialbot
[REVIEW]: tukey_hsd: An Accurate Implementation of the Tukey Honestly Significant Difference Test in Python
**Submitting author:** @mdhaber (<a href="http://orcid.org/0000-0003-4806-3601">Matt Haberland</a>)
**Repository:** https://github.com/mdhaber/scipy/
**Branch with paper.md** (empty if default branch): joss_studentized_range
**Version:** 1.8.0
