### 2022-05-09 #4383 @mikldk
@mcavs, @acolum: Thanks for agreeing to review. Please carry out your review in this issue by first creating a checklist (`@editorialbot generate my checklist`) and then updating it as the review proceeds. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.
