### 2022-05-09 #4159 @mdhaber
@mikldk Since the relevant code is integrated into the (very large) SciPy library, would it be appropriate to copy https://github.com/openjournals/joss-reviews/issues/4159#issuecomment-1039235888 and https://github.com/openjournals/joss-reviews/issues/4159#issuecomment-1039328191 over to the new issue to help with review?
