### 2022-05-09 #4383 @acolum
@mikldk Thank you, and I have a quick question. I volunteered to review submissions this year for the SciPy conference - would this be considered a COI according to the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md)?
