### 2022-05-12 #4383 @mikldk
> @mikldk Thank you, and I have a quick question. I volunteered to review submissions earlier this year for the SciPy conference - would this be considered a COI according to the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md)?

@acolum I think it is fine that this potential COI is now recorded, and I think it can be waived if you think you are able to make an impartial assessment of the work?
