### 2022-05-12 #4159 @mikldk
> @mikldk Since the relevant code is integrated into the (very large) SciPy library, would it be appropriate to copy [#4159 (comment)](https://github.com/openjournals/joss-reviews/issues/4159#issuecomment-1039235888) and [#4159 (comment)](https://github.com/openjournals/joss-reviews/issues/4159#issuecomment-1039328191) over to the new issue to help with review?

Good idea, thanks. I have made [this comment](https://github.com/openjournals/joss-reviews/issues/4383#issuecomment-1124743074) in the review. You are welcome to insert your answer there, too, if you like.
