### 2022-05-18 mdhaber/scipy#78 @acolum
JOSS Review Suggestions
Hi @mdhaber,

Nice work on this submission. I have a few small suggestions for improving your paper:

Under *Summary*:
1. Please change "in hope of" to "in **the** hope of."
2. Please change "Carefully designed tests" to "Carefully-designed tests."
3. Please rephrase' "p-hacking", producing' to ' "p-hacking" by producing' for greater clarity.

Under *Statement of need*:
1. Please change "cumulative density function (CDF)" to "cumulative **distribution** function (CDF)."
2. On line 37, "cumulative distribution function" could be replaced with "CDF."

Thanks,
Alyssa

For Reference: [Link to Review](https://github.com/openjournals/joss-reviews/issues/4383)
