### 2022-05-18 c6b7f53
Respond to suggestions in mdhaber/scipy#78
