### 2022-05-18 mdhaber/scipy#78 @mdhaber
Thanks @acolum! I opened gh-79 so that the changes are easy to track.
