### 2022-05-18 mdhaber/scipy#78 @mdhaber
Thanks again @acolum. It sounds like gh-79 addressed this, but if there were any remaining concerns, please let me know.
