### 2022-05-18 mdhaber/scipy#78 @acolum
@mdhaber, thanks for making those changes!

Unfortunately, I noticed that one instance of "CDF" remained in the paper.md file even after the pull request was merged.
