### 2022-05-18 mdhaber/scipy#78 @mdhaber
Oops! Fixed in d6385bcef595eadfbef815b62fe53670199f7716. (Looks like my IDE also automatically added a newline at the end of the file, but I don't think that affected the way the file renders.)
