### 2022-05-18 mdhaber/scipy#78 @acolum
Awesome, thanks! I'll continue my review back in the original JOSS issue.
