### 2022-05-23 mdhaber/scipy#82 @mcavs
JOSS Review
Dear @mdhaber,

Your paper is well-written and pretty understandable. But I think it needs a minor addition:
* You mentioned that "A through assessment of the methods, accuracy, and speed of the underlying calculations is available in Wallen et al. (2021a)" in the abstract of the paper. These comparisons may be given as a summary under a new section by highlighting the speed, and the other properties of the proposed version of HSD test, after Sec 2. I think it is difficult to follow on the referenced paper.

* The citation of Wallen et al. (2021a) which is in abstract should be out of the brackets.

Thanks,
Mustafa

For Reference: [Link to review](https://github.com/openjournals/joss-reviews/issues/4383)
