### 2022-05-24 mdhaber/scipy#82 @mdhaber
Thanks @mcavs. I will submit a PR for your review shorty. But to clarify:

> The citation of Wallen et al. (2021a) which is in abstract should be out of the brackets.

Do you mean that the citation in the summary section should not be surrounded by parentheses? That is,
"the underlying calculations is available in (Wallan et al., 2021a),"
should appear instead as
"the underlying calculations is available in Wallan et al., 2021a,"?
If so, I'm not sure how to control how citations appear: in the source, this appears as `[@StudentizedRangeSciPy]`, but it renders as "(Wallan et al., 2021a)". Do you know how to control this? If not, this might be a question for the editor. In some citation styles (e.g. IEEE), citations can be used as though they were a noun in the sentence rather than as a parenthetical note, in which case the citation might be OK as it appears.
