### 2022-05-24 mdhaber/scipy#82 @mcavs
Hi @mdhaber,

I am sorry, it's my fault, your usage is correct according to the journal's citation format.
I withdraw my suggestion regarding this change.

Thanks for asking!
