### 2022-05-24 d21456e
Add Performance section
