### 2022-05-24 mdhaber/scipy#82 @mdhaber
Thanks @mcavs. In that case, does gh-83 resolve this issue?
