### 2022-06-09 #4383 @mikldk
@mcavs: Can you please update us on your status? All items are checked so are you now recommending an accept? I see mdhaber/scipy#82 is still not closed.
