### 2022-06-09 mdhaber/scipy#82 @mcavs
I think it is ready to flight, @mdhaber!
