### 2022-06-09 #4383 @mcavs
A sufficient [revision](https://github.com/mdhaber/scipy/pull/83) has been made by the author, and I recommend accepting @mikldk.
