### 2022-06-10 #4383 @mdhaber
@mcavs https://github.com/mdhaber/scipy/issues/83 has not been merged. Would that resolve your concerns expressed in https://github.com/mdhaber/scipy/issues/82? If so, could you indicate your approval in https://github.com/mdhaber/scipy/issues/83? Thanks!
