+++ scipy/stats/_stats_py.py
@@ -156,6 +156,7 @@ def _shape_with_dropped_axis(a, axis):
 
     Examples
     --------
+    >>> import numpy as np
     >>> a = np.zeros((3, 5, 2))
     >>> _shape_with_dropped_axis(a, 1)
     (3, 2)
@@ -208,6 +209,7 @@ def _broadcast_shapes_with_dropped_axis(a, b, axis):
 
     Examples
     --------
+    >>> import numpy as np
     >>> a = np.zeros((5, 2, 1))
     >>> b = np.zeros((1, 9, 3))
     >>> _broadcast_shapes_with_dropped_axis(a, b, 1)
@@ -587,6 +589,7 @@ def mode(a, axis=0, nan_policy='propagate'):
 
     Examples
     --------
+    >>> import numpy as np
     >>> a = np.array([[6, 8, 3, 0],
     ...               [3, 2, 1, 7],
     ...               [8, 1, 8, 4],
@@ -723,6 +726,7 @@ def tmean(a, limits=None, inclusive=(True, True), axis=None):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tmean(x)
@@ -776,6 +780,7 @@ def tvar(a, limits=None, inclusive=(True, True), axis=0, ddof=1):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tvar(x)
@@ -829,6 +834,7 @@ def tmin(a, lowerlimit=None, axis=0, inclusive=True, nan_policy='propagate'):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tmin(x)
@@ -890,6 +896,7 @@ def tmax(a, upperlimit=None, axis=0, inclusive=True, nan_policy='propagate'):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tmax(x)
@@ -953,6 +960,7 @@ def tstd(a, limits=None, inclusive=(True, True), axis=0, ddof=1):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tstd(x)
@@ -1001,6 +1009,7 @@ def tsem(a, limits=None, inclusive=(True, True), axis=0, ddof=1):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> x = np.arange(20)
     >>> stats.tsem(x)
@@ -1390,6 +1399,7 @@ def kurtosis(a, axis=0, fisher=True, bias=True, nan_policy='propagate'):
     >>> import scipy.stats as stats
     >>> from scipy.stats import kurtosis
 
+    >>> import numpy as np
     >>> x = np.linspace(-5, 5, 100)
     >>> ax = plt.subplot()
     >>> distnames = ['laplace', 'norm', 'uniform']
@@ -1496,6 +1506,7 @@ def describe(a, axis=0, ddof=1, bias=True, nan_policy='propagate'):
 
     Examples
     --------
+    >>> import numpy as np
     >>> from scipy import stats
     >>> a = np.arange(10)
     >>> stats.describe(a)
@@ -1716,6 +1727,7 @@ def kurtosistest(a, axis=0, nan_policy='propagate', alternative='two-sided'):
     >>> kurtosistest(list(range(20)), alternative='greater')
     KurtosistestResult(statistic=-1.7058104152122062, pvalue=0.9559783083373583)
 
+    >>> import numpy as np
     >>> rng = np.random.default_rng()
[…]
