### 2022-06-15 e33072c
MAINT: add import numpy as np to more files
