### 2022-06-15 #4383 @mdhaber
Thanks @mcavs, @acolum!

@mikldk Thanks! I'll do the final checks.

ALso, the software reported here is part of SciPy, which is a very complex library. Do you have any recommendations about the preferred repo state? The version of the repo I added the paper to was branched from SciPy `main` sometime between releases. SciPy 1.9 should be released soon - should I merge the released version of 1.9 into this branch, and would it be OK to leave the branch as-is? Should I clean up the commit history of the paper, or leave that alone? In any case, I wouldn't recommend that users try to use my release of the library; they really should be referred to an official release (which also contains the code reported in the paper). Is that possible?
