### 2022-06-23 4f76c97
Fix redundancy in introduction
