### 2022-06-26 e29921b
DOC: optimize: release notes update
