### 2022-06-30 #4383 @mdhaber
@mikldk Thanks! I think those questions are resolved from my perspective. If the release/DOI look OK to JOSS, I think we're set.
