### 2022-06-30 #4383 @mikldk
> @mikldk Thanks! I think those questions are resolved from my perspective. If the release/DOI look OK to JOSS, I think we're set.

@mdhaber That sounds good. Can you please briefly describe how/why it is resolved?
