### 2022-06-30 #4383 @mdhaber
Sure. I had the following questions and made the decisions described below.

> would it be OK to leave the branch as-is?

I decided "yes" because the branch is supposed to released be in the state it was reviewed in, so it would not be appropriate to pull in other commits.

> Should I clean up the commit history of the paper...?

I decided "no" because the commit history should reflect that changes were made due to reviewer feedback, so I should not clean that up.

> they really should be referred to an official release (which also contains the code reported in the paper). Is that possible?

Yes, it is; the release notes of my release link to the SciPy official release page.
