### 2022-07-05 #4383 @danielskatz
@mdhaber - I'm the AEiC on duty this week, and I will complete the processing of this paper.  I've suggested a small change in https://github.com/mdhaber/scipy/pull/86 - if this is ok, please merge it and we can proceed.
